Wrote convolve function using fftpack from scipy
Many iterations involving different fft libraries and issues with padding
finally resolved by putting the padded shape size in the fft function
this means that the resulting image will have a pixel on all sides of zeros
in the returned function at the moment. This will be added in later commits.
Also added pass to other functions to make script running work.