Added code to remove padding from convolved. Test suite working for convolve.
Test suite now checks that my convolve and the scipy convolve are the same to 6 decimal places, whether the size of out is the same as input and if errors are raised when even kernels are given.

In hybrid_image.py added function to check more arbitrary odd numbers, and supported types. Padding is removed from the array convolved to return image without borders. The gaussian kernel creator has odd and type checking and has the beginnings of workings.