Convolved works for 3 dimension, high and low pass filters written
3d convolving works using recursion. low pass calls gaussian kernel
to then convolve over. High pass calls low pass and subtracts low
frequencies from original image. Initial adding of two images together
to create hybrid written not tested. Custom imshow function made so that
outputs of type float not limited to between 0 and 1 can be shown using imshow
(this converts to uint8 whilst leaving the output untouched) and most docstringshave been written.

Testing suite now includes a test to check that an n-dimensional image returns
an n-dimensional image of the same shape. It also checks that nothing greater than three dimensions is accepted and that the output from a known convolve copied
into 3 dimensions is the same as doing it in 3 dimensions from the start. Test
suite is currently OK.